# backend/app/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Uni-Chat API
  description: |
    Uni-Chat is a full-stack AI chat application using OpenRouter for multi-model access.

    Features include:
    - Real-time streaming chat
    - Image generation
    - Workflow editor
    - Arena mode for side-by-side AI comparison

    ## Authentication
    All protected endpoints require a JWT Bearer token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "error": "Error message description"
    }
    ```
  version: 1.0.0
  contact:
    name: Uni-Chat Support
servers:
  - url: /api
    description: API base path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          type: string
          enum: [user, admin]
        avatar_url:
          type: string
          nullable: true
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        profile:
          type: object
          properties:
            display_name:
              type: string
            avatar_url:
              type: string
              nullable: true
            bio:
              type: string
        settings:
          type: object
        usage:
          type: object
          properties:
            messages_sent:
              type: integer
            tokens_used:
              type: integer
            tokens_limit:
              type: integer
        created_at:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Token expiration in seconds
          example: 900
        user:
          $ref: '#/components/schemas/User'
...
    TokenRefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 900
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login and get access tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Missing credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account suspended
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      operationId: refreshToken
      security:
        - BearerAuth: []
      description: Requires refresh token (not access token) in Authorization header
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: Logout and revoke token
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Auth
      summary: Get current user information
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/password:
    put:
      tags:
        - Auth
      summary: Change password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - current_password
                - new_password
              properties:
                current_password:
                  type: string
                  format: password
                new_password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Conversations endpoints
...